Added Linting and formating on commit
