+++ package.json
@@ -12,7 +12,15 @@
     "build-only": "vite build",
     "type-check": "vue-tsc --build --force",
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
-    "format": "prettier --write src/"
+    "format": "prettier --write src/",
+    "prepare": "husky",
+    "pre-commit": "lint-staged"
+  },
+  "lint-staged": {
+    "src/**/*.{ts,tsx,js,jsx,vue}": [
+      "npm run format",
+      "npm run lint"
+    ]
   },
   "dependencies": {
     "@shadcn/ui": "^0.0.4",
@@ -20,6 +28,8 @@
     "axios": "^1.7.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "husky": "^9.1.6",
+    "lint-staged": "^15.2.10",
     "lucide-vue-next": "^0.447.0",
     "pinia": "^2.1.7",
     "radix-vue": "^1.9.6",
@@ -43,7 +53,7 @@
     "eslint-plugin-vue": "^9.23.0",
     "jsdom": "^24.1.0",
     "npm-run-all2": "^6.2.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.3",
     "tailwindcss": "^3.4.13",
     "typescript": "~5.4.0",
     "vite": "^5.3.1",
