+++ src/components/LoginView.vue
@@ -3,9 +3,9 @@ import { ref } from 'vue'
 import { useUserStore } from '@/stores/userStore'
 import { useRouter } from 'vue-router'
 
-import Button from '@/components/ui/button/Button.vue'
+import Button from '@/components/ui/button/BaseButton.vue'
 import Input from '@/components/ui/input/Input.vue'
-import Label from '@/components/ui/label/Label.vue'
+import Label from '@/components/ui/label/BaseLabel.vue'
 
 const userStore = useUserStore()
 const router = useRouter()
@@ -13,9 +13,8 @@ const email = ref('')
 const password = ref('')
 const error = ref('')
 
-const handleSignup = async () => {
+const handleLogin = async () => {
   try {
-    await userStore.signup(email.value, password.value)
     await userStore.login(email.value, password.value)
     router.push('/dashboard')
   } catch (err) {
@@ -23,11 +22,10 @@ const handleSignup = async () => {
   }
 }
 </script>
-
 <template>
   <div class="max-w-md mx-auto p-4">
-    <h2 class="text-2xl font-semibold mb-4">Sign Up</h2>
-    <form @submit.prevent="handleSignup">
+    <h2 class="text-2xl font-semibold mb-4">Login</h2>
+    <form @submit.prevent="handleLogin">
       <div class="mb-3">
         <Label for="email">Email</Label>
         <Input id="email" v-model="email" type="email" required placeholder="Enter your email" />
@@ -42,7 +40,7 @@ const handleSignup = async () => {
           placeholder="Enter your password"
         />
       </div>
-      <Button type="submit" class="w-full">Sign Up</Button>
+      <Button type="submit" class="w-full">Login</Button>
     </form>
     <p v-if="error" class="text-red-500 mt-2">{{ error }}</p>
   </div>
