+++ src/components/__tests__/HelloWorld.spec.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'vitest'
+import { describe, it } from 'vitest'
 
 describe('HelloWorld', () => {
   it('renders properly', () => {
