+++ src/components/ui/input/index.ts
@@ -1,1 +1,1 @@
-export { default as Input } from './Input.vue'
+export { default as Input } from './BaseInput.vue'
