+++ src/components/ui/label/BaseLabel.vue
@@ -18,7 +18,7 @@ const delegatedProps = computed(() => {
     :class="
       cn(
         'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
-        props.class,
+        props.class
       )
     "
   >
