+++ src/components/ui/label/index.ts
@@ -1,1 +1,1 @@
-export { default as Label } from './Label.vue'
+export { default as Label } from './BaseLabel.vue'
